Improve detection of distribution family
Use "ID" field in /usr/lib/os-release to find the distribution type.
This will be used whenever "ID_LIKE" is not defined.

Cleanroom wants to know the distribution family, not the actual
distribution, so ID_LIKE is the better choice, but some distributions
that are stand-alone do not bother to set that field.

Closes: #10