Scan for qemu-img and qemu-nbd
The commands are not used via the Binary Manager that does the
scanning right now, but at least the commands get warned about
when building a system for the first time.

Next up: Actually use the detected commands in firestarter which
currently does not use the binary manager at all.

Closes: #13