NbdDevice: Skip unsuitable nbd devices
Skip /dev/nbdX if that is not a block device or if it is already
connected to something. This should make the nbd device handling
more robust in the face of people actually using them;-)

Get the number of possible nbd devices out of sys (with a fallback
to 32), so that there is no need to loop over too many device nodes.