Command: Allow for extra files/binaries/modules/hooks in create_initrd.py
Evaluate the MKINITCPIO_EXTRA_(FILES|BINARIES|MODULES|HOOKS)
substitutions and add those to the generated initrd.

Adapt examples to not include mkinitcpio explictily anymore. This means
that grep is also not needed anymore: It was there as a hack to avoid it
getting removed -- even when explicitly installed later -- when
mkinitcpio gets removed again.