pkg_usbguard: Make daemon config persistent
Move all settings into /var/etc/usbguard to make them persistent.

This allows to start with a default allow-everything policy in place
(the new default) that can be tightened later.

This makes it possible to use pkg_usbguard on a machine with
USB keyboard/mouse without locking yourself out entirely:-)