Clrm: Remove TPM support from initrd
This was a quick hack I did for one of my old laptops that should
probably never have made it out into the wild.

It did dump the TPM registers, pass all of them through md5sum and
then turned that into words to give an idea of the code that had
been loaded so far. It does not really improve security, and was
more informational than anythig else.